fix for correct structure size & alignment
git-svn-id: http://miranda.googlecode.com/svn/trunk@6541 83df23c2-51fd-11de-bb24-2d5e9b4af41d

+++ miranda/src/modules/utils/resizer.c
@@ -30,7 +30,7 @@ typedef struct {
 	short  y;
 	short  cx;
 	short  cy;
-	WORD   id;
+	DWORD  id;
 } START_OF_DLGITEMTEMPLATEEX;
 
 typedef struct {
